% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-dataset-urls.R
\docType{data}
\name{catalog_root_urls}
\alias{catalog_root_urls}
\title{Source data root URL path}
\format{
An object of class \code{list} of length 4.
}
\usage{
catalog_root_urls
}
\description{
Source data root URL path
}
\keyword{datasets}
