% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-dataset-urls.R
\docType{data}
\name{ccia_access_types}
\alias{ccia_access_types}
\title{Dataset access types}
\format{
An object of class \code{list} of length 8.
}
\usage{
ccia_access_types
}
\description{
Dataset access types
}
\keyword{datasets}
