% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccia_add_netcdf_subset_query.R
\name{ccia_add_netcdf_subset_query}
\alias{ccia_add_netcdf_subset_query}
\title{Add query parameters of CCIA NetCDF Subset Service to a request URL}
\usage{
ccia_add_netcdf_subset_query(
  request,
  vars,
  bbox,
  lat,
  lon,
  datetime_start,
  datetime_end,
  datetime_step
)
}
\arguments{
\item{request}{httr2 request}

\item{vars}{dataset variables to download. Use 'all' for all available.}

\item{bbox}{number, vector of length 4: left,bottom,right,top | min Longitude , min Latitude , max Longitude , max Latitude}

\item{lat, }{number, latitude}

\item{lon}{number, longitude}

\item{datetime_start}{date, can be character, Date or POSIX*}

\item{datetime_end}{date, can be character, Date or POSIX*}

\item{datetime_step}{number, days between date_start and date_end to download}
}
\value{
character, path to temporary file that is downloaded
}
\description{
Add query parameters to a CCIA request URL. Note that returned data is dependent
on what is available and might not exactly match the request e.g. where the time
slices are centered monthly or yearly.
}
