% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccia_create_dataset_request_url.R
\name{ccia_create_dataset_request_url}
\alias{ccia_create_dataset_request_url}
\title{Create CCIA dataset URL request}
\usage{
ccia_create_dataset_request_url(
  dataset_relative_url,
  access_type = c("NetcdfSubset", "OPENDAP", "HTTPServer", "WMS", "WCS", "NCML", "UDDC",
    "ISO")
)
}
\arguments{
\item{dataset_relative_url}{relative URL from the dataset download page.
Loaded into R session with \code{\link[=ccia_dataset_urls]{ccia_dataset_urls()}}}

\item{access_type}{dataset type to download.}
}
\value{
An HTTP response: an S3 list with class httr2_request.
}
\description{
Start the process of building a CCIA dataset URL request. This function takes
a relative URL (see \code{\link[=ccia_dataset_urls]{ccia_dataset_urls()}}) and a given
access type and creates a \code{\link[httr2:request]{httr2::request()}}
}
\details{
There are two base URLs and internal URL paths depending on the dataset type (Projected change
or Plausible future climate data and Threshold datasets). This function builds
the URL given the dataset and access type https://\if{html}{\out{<dataset-path>}}/\if{html}{\out{<access-path>}}/\if{html}{\out{<dataset-path2>}}
}
