% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccia_extract_dataset_xml_data.R
\name{ccia_extract_dataset_xml_data}
\alias{ccia_extract_dataset_xml_data}
\title{Extract variable and timespan data from data description XML}
\usage{
ccia_extract_dataset_xml_data(xml)
}
\arguments{
\item{xml}{xml2::xml_document}
}
\value{
a list
}
\description{
Extract \code{grid} and \code{TimeSpan} elements from data description XML into a list
}
