% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ccia_get_dataset_netcdf_subset_description_xml.R
\name{ccia_get_dataset_netcdf_subset_description_xml}
\alias{ccia_get_dataset_netcdf_subset_description_xml}
\title{Get dataset description XML}
\usage{
ccia_get_dataset_netcdf_subset_description_xml(request)
}
\arguments{
\item{request}{httr2 request with 'NetcdfSubset' access type}
}
\value{
an \code{xml2::xml_document}
}
\description{
Get data description XML from a build dataset URL (before query parameters added)
}
