% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccia_get_dataset_netcdf_subset_info.R
\name{ccia_get_dataset_netcdf_subset_info}
\alias{ccia_get_dataset_netcdf_subset_info}
\title{Get dataset description in easy to view and handle manner}
\usage{
ccia_get_dataset_netcdf_subset_info(request)
}
\arguments{
\item{request}{httr2 request with 'NetcdfSubset' access type}
}
\value{
a list
}
\description{
Obtain the dataset description XML and wrangle into a list. Works only
for NetcdfSubset URLs.
}
\details{
Uses \code{\link[=ccia_get_dataset_netcdf_subset_description_xml]{ccia_get_dataset_netcdf_subset_description_xml()}}
to obtain the XML and then \code{\link[=ccia_extract_dataset_xml_data]{ccia_extract_dataset_xml_data()}}
}
