% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccia_perform_query.R
\name{ccia_perform_query}
\alias{ccia_perform_query}
\title{Perform a request from a built CCIA query}
\usage{
ccia_perform_query(request, destfile)
}
\arguments{
\item{request}{httr2 request with query parameters}

\item{destfile}{where the response body should be written}
}
\value{
filepath character vector
}
\description{
Perform a request from a built CCIA query
}
