% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generic.R
\name{check_bbox}
\alias{check_bbox}
\title{Check a given bounding box list values are correct}
\usage{
check_bbox(x)
}
\arguments{
\item{x}{a list with 4 named elements}
}
\value{
nothing
}
\description{
Ensure the bbox values are in the correct order
}
