% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-dataset-urls.R
\name{dataset_nodeset_to_df}
\alias{dataset_nodeset_to_df}
\title{Extract attributes from dataset nodes}
\usage{
dataset_nodeset_to_df(xml_nodeset)
}
\arguments{
\item{xml_nodeset}{an xml_nodeset}
}
\description{
Extract attributes from dataset nodes
}
