% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generic.R
\name{format_datetime}
\alias{format_datetime}
\title{Format a datetime for a NetCDF query}
\usage{
format_datetime(x)
}
\arguments{
\item{x}{a character, or datetime (Date, as.POSIX*)}
}
\value{
formatted character value
}
\description{
Format a datetime for a NetCDF query
}
