% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-dataset-urls.R
\name{get_dataset_links}
\alias{get_dataset_links}
\title{Recurse through a catalog.xml to find dataset links}
\usage{
get_dataset_links(url, .test = FALSE)
}
\arguments{
\item{url}{the top level URL to start searching}

\item{.test}{boolean, used in testing to only recurse 2 levels at maximum}
}
\description{
Simulates clicking through the catalog as viewed in the brower, but by examing
web links.
}
\examples{
\dontrun{

url_root <-
  file.path('https://dapds00.nci.org.au/thredds/catalog',
  'ua6_4/CMIP5/derived/Collections/Projected_Change_Data')

datasets_many <- get_dataset_links(url = url_root)

url_root <-
  file.path('https://dapds00.nci.org.au/thredds/catalog',
  'ua6_4/CMIP5/derived/Collections/Projected_Change_Data/Drought')

datasets_drought <- get_dataset_links(url = url_root)
}
}
