% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-dataset-urls.R
\name{save_data_urls_to_json}
\alias{save_data_urls_to_json}
\title{Recurse through the CSIRO catalog to obtain all dataset URLs}
\usage{
save_data_urls_to_json(destfile = "inst/dataset-urls.json")
}
\arguments{
\item{destfile}{where to save the JSON}
}
\description{
Used to create a catalog JSON
}
